requirements.txt
cycler==0.11.0
easydict==1.10
fonttools==4.38.0
joblib==1.2.0
numpy==1.21.6
opencv-python-headless==4.7.0.72
packaging==23.0
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
scipy==1.7.3
six==1.16.0
threadpoolctl==3.1.0
typing_extensions==4.5.0
fonttools==4.38.0
tqdm==4.65.0
albumentations==1.3.0